Develop

+++ src/main/java/com/wallet/mono/service/serviceImpl/AccountServiceImpl.java
@@ -7,13 +7,17 @@ import com.wallet.mono.domain.dto.FavoriteRequest;
 import com.wallet.mono.domain.mapper.AccountRequestMapper;
 import com.wallet.mono.domain.mapper.AccountResponseMapper;
 import com.wallet.mono.domain.model.Account;
+import com.wallet.mono.domain.model.Transaction;
 import com.wallet.mono.exception.AccountAlreadyExistsException;
 import com.wallet.mono.exception.AccountNotFoundException;
+import com.wallet.mono.exception.UnableToDeleteFavoriteAccount;
 import com.wallet.mono.exception.UserNotFoundException;
 import com.wallet.mono.repository.AccountRepository;
+import com.wallet.mono.repository.TransactionRepository;
 import com.wallet.mono.service.AccountService;
 import com.wallet.mono.service.UserService;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -29,6 +33,7 @@ public class AccountServiceImpl implements AccountService {
     private final AccountRepository accountRepository;
     private final AccountRequestMapper accountRequestMapper;
     private final AccountResponseMapper accountResponseMapper;
+    private final TransactionRepository transactionRepository;
 
     @Override
     public void saveAccount(AccountRequest accountRequest) throws Exception {
@@ -43,6 +48,11 @@ public class AccountServiceImpl implements AccountService {
         }
 
         Account account = accountRequestMapper.mapToAccount(accountRequest);
+
+        if (accounts.isEmpty()) {
+            account.setFavorite(true);
+        }
+
         accountRepository.save(account);
 
     }
@@ -138,7 +148,19 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public void deleteAccount(FavoriteRequest favoriteRequest) throws Exception {
-        getAccountDetails(favoriteRequest.getAccountId(), favoriteRequest.getUserId());
+        AccountResponse accountResponse = getAccountDetails(favoriteRequest.getAccountId(), favoriteRequest.getUserId());
+
+        if (accountResponse.isFavorite()) {
+            throw new UnableToDeleteFavoriteAccount();
+        }
+
+        List<Transaction> transactions = transactionRepository.findByAccount_AccountId(favoriteRequest.getAccountId(), Pageable.unpaged()).getContent();
+
+        if (!transactions.isEmpty()) {
+            List<Integer> transactionIds = transactions.stream().map(Transaction::getTransactionId).toList();
+            transactionRepository.deleteAllById(transactionIds);
+        }
+
         accountRepository.deleteById(favoriteRequest.getAccountId());
     }
